refactor(http): split websocket bridge into qemu-ws
